+++ extensions/amp-story-share-menu/0.1/test/test-amp-story-share-menu.js
@@ -1,0 +1,106 @@
+import {Keys_Enum} from '#core/constants/key-codes';
+
+import {Services} from '#service';
+
+import {registerServiceBuilder} from '../../../../src/service-helpers';
+import {
+  Action,
+  AmpStoryStoreService,
+  StateProperty,
+} from '../../../amp-story/1.0/amp-story-store-service';
+import {AmpStoryShareMenu, VISIBLE_CLASS} from '../amp-story-share-menu';
+
+describes.realWin('amp-story-share-menu', {amp: true}, (env) => {
+  let hostEl;
+  let shareMenu;
+  let ampStory;
+  let storeService;
+  let win;
+  let installExtensionForDoc;
+
+  beforeEach(() => {
+    win = env.win;
+    storeService = new AmpStoryStoreService(win);
+    registerServiceBuilder(win, 'story-store', function () {
+      return storeService;
+    });
+
+    // Making sure the vsync tasks run synchronously.
+    env.sandbox.stub(Services, 'vsyncFor').returns({
+      mutate: (fn) => fn(),
+      run: (vsyncTaskSpec, vsyncState) => {
+        vsyncTaskSpec.measure(vsyncState);
+        vsyncTaskSpec.mutate(vsyncState);
+      },
+    });
+
+    installExtensionForDoc = env.sandbox.spy();
+
+    env.sandbox.stub(Services, 'extensionsFor').returns({
+      installExtensionForDoc,
+    });
+
+    hostEl = win.document.createElement('div');
+    ampStory = win.document.createElement('amp-story');
+    win.document.body.appendChild(ampStory);
+    ampStory.appendChild(hostEl);
+    shareMenu = new AmpStoryShareMenu(hostEl);
+  });
+
+  it('should build the sharing menu', () => {
+    shareMenu.build();
+    expect(shareMenu.element_).to.exist;
+  });
+
+  it('should show the share menu on store property update', () => {
+    shareMenu.build();
+
+    storeService.dispatch(Action.TOGGLE_SHARE_MENU, true);
+
+    expect(shareMenu.element_).to.have.class(VISIBLE_CLASS);
+  });
+
+  it('should hide the share menu on click on the overlay', () => {
+    shareMenu.build();
+
+    storeService.dispatch(Action.TOGGLE_SHARE_MENU, true);
+    shareMenu.element_.dispatchEvent(new Event('click'));
+
+    expect(shareMenu.element_).not.to.have.class(VISIBLE_CLASS);
+    expect(storeService.get(StateProperty.SHARE_MENU_STATE)).to.be.false;
+  });
+
+  it('should not hide the share menu on click on the widget container', () => {
+    shareMenu.build();
+
+    storeService.dispatch(Action.TOGGLE_SHARE_MENU, true);
+    hostEl
+      .querySelector('.i-amphtml-story-share-menu-container')
+      .dispatchEvent(new Event('click'));
+
+    expect(shareMenu.element_).to.have.class(VISIBLE_CLASS);
+  });
+
+  it('should hide the share menu on escape key press', () => {
+    shareMenu.build();
+
+    const clickCallbackSpy = env.sandbox.spy();
+    win.addEventListener('keyup', clickCallbackSpy);
+
+    storeService.dispatch(Action.TOGGLE_SHARE_MENU, true);
+    // Create escape keyup event.
+    const keyupEvent = new Event('keyup');
+    keyupEvent.keyCode = Keys_Enum.ESCAPE;
+    win.dispatchEvent(keyupEvent);
+
+    expect(clickCallbackSpy).to.have.been.calledOnce;
+  });
+
+  it('should load the amp-social-share extension', () => {
+    shareMenu.build();
+
+    expect(
+      installExtensionForDoc.withArgs(env.sandbox.match.any, 'amp-social-share')
+    ).to.have.been.called;
+  });
+});
